+++ Assets/AppCenter/Plugins/AppCenterSDK/Distribute/Shared/ReleaseAvailableCallback.cs
@@ -7,4 +7,9 @@ namespace Microsoft.AppCenter.Unity.Distribute
     /// Release available callback.
     /// </summary>
     public delegate bool ReleaseAvailableCallback(ReleaseDetails releaseDetails);
+
+    /// <summary>
+    /// App will exit callback.
+    /// </summary>
+    public delegate void WillExitAppCallback();
 }