+++ Assets/Puppet/PuppetReleaseHandler.cs
@@ -15,6 +15,7 @@ public class PuppetReleaseHandler : MonoBehaviour
     void Awake()
     {
         Distribute.ReleaseAvailable = OnReleaseAvailable;
+        Distribute.WillExitApp = OnWillExitApp;
         IsDialogCustom = PlayerPrefs.GetInt(PuppetAppCenter.FlagCustomDialog, 0) == 1;
     }
 
@@ -28,6 +29,14 @@ public class PuppetReleaseHandler : MonoBehaviour
                 return true;
             }
             return false;
+        }
+    }
+
+    void OnWillExitApp()
+    {
+        lock (_releaseLock)
+        {
+            Debug.Log("WillExitApp callback invoked.");
         }
     }